+++ mmpose/datasets/datasets/top_down/topdown_coco_wholebody_dataset.py
@@ -6,7 +6,7 @@ import numpy as np
 from mmcv import Config
 from xtcocotools.cocoeval import COCOeval
 
-from mmpose.datasets.builder import DATASETS
+from ...builder import DATASETS
 from .topdown_coco_dataset import TopDownCocoDataset
