--- tests/test_datasets/test_wholebody_dataset.py
@@ -1,114 +1,0 @@
-import copy
-import tempfile
-from unittest.mock import MagicMock
-
-import numpy as np
-import pytest
-from numpy.testing import assert_almost_equal
-
-from mmpose.datasets import DATASETS
-
-
-def convert_db_to_output(db, batch_size=2):
-    outputs = []
-    len_db = len(db)
-    for i in range(0, len_db, batch_size):
-        keypoints = np.stack([
-            db[j]['joints_3d'].reshape((-1, 3))
-            for j in range(i, min(i + batch_size, len_db))
-        ])
-        image_paths = [
-            db[j]['image_file'] for j in range(i, min(i + batch_size, len_db))
-        ]
-        bbox_ids = [j for j in range(i, min(i + batch_size, len_db))]
-        box = np.stack(
-            np.array([
-                db[j]['center'][0], db[j]['center'][1], db[j]['scale'][0],
-                db[j]['scale'][1], db[j]['scale'][0] * db[j]['scale'][1] *
-                200 * 200, 1.0
-            ],
-                     dtype=np.float32)
-            for j in range(i, min(i + batch_size, len_db)))
-
-        output = {}
-        output['preds'] = keypoints
-        output['boxes'] = box
-        output['image_paths'] = image_paths
-        output['output_heatmap'] = None
-        output['bbox_ids'] = bbox_ids
-
-        outputs.append(output)
-
-    return outputs
-
-
-def test_top_down_COCO_wholebody_dataset():
-    dataset = 'TopDownCocoWholeBodyDataset'
-    # test COCO datasets
-    dataset_class = DATASETS.get(dataset)
-    dataset_class.load_annotations = MagicMock()
-    dataset_class.coco = MagicMock()
-
-    channel_cfg = dict(
-        num_output_channels=133,
-        dataset_joints=133,
-        dataset_channel=[
-            list(range(133)),
-        ],
-        inference_channel=list(range(133)))
-
-    data_cfg = dict(
-        image_size=[192, 256],
-        heatmap_size=[48, 64],
-        num_output_channels=channel_cfg['num_output_channels'],
-        num_joints=channel_cfg['dataset_joints'],
-        dataset_channel=channel_cfg['dataset_channel'],
-        inference_channel=channel_cfg['inference_channel'],
-        soft_nms=False,
-        nms_thr=1.0,
-        oks_thr=0.9,
-        vis_thr=0.2,
-        use_gt_bbox=True,
-        det_bbox_thr=0.0,
-        bbox_file='tests/data/coco/test_coco_det_AP_H_56.json',
-    )
-    # Test det bbox
-    data_cfg_copy = copy.deepcopy(data_cfg)
-    data_cfg_copy['use_gt_bbox'] = False
-    _ = dataset_class(
-        ann_file='tests/data/coco/test_coco_wholebody.json',
-        img_prefix='tests/data/coco/',
-        data_cfg=data_cfg_copy,
-        pipeline=[],
-        test_mode=True)
-
-    _ = dataset_class(
-        ann_file='tests/data/coco/test_coco_wholebody.json',
-        img_prefix='tests/data/coco/',
-        data_cfg=data_cfg_copy,
-        pipeline=[],
-        test_mode=False)
-
-    # Test gt bbox
-    custom_dataset = dataset_class(
-        ann_file='tests/data/coco/test_coco_wholebody.json',
-        img_prefix='tests/data/coco/',
-        data_cfg=data_cfg,
-        pipeline=[],
-        test_mode=True)
-
-    assert custom_dataset.test_mode is True
-    assert custom_dataset.dataset_name == 'coco_wholebody'
-
-    image_id = 785
-    assert image_id in custom_dataset.img_ids
-    assert len(custom_dataset.img_ids) == 4
-    _ = custom_dataset[0]
-
-    outputs = convert_db_to_output(custom_dataset.db)
-    with tempfile.TemporaryDirectory() as tmpdir:
-        infos = custom_dataset.evaluate(outputs, tmpdir, 'mAP')
-        assert_almost_equal(infos['AP'], 1.0)
-
-        with pytest.raises(KeyError):
-            _ = custom_dataset.evaluate(outputs, tmpdir, 'PCK')
